mock data refactor
