+++ Demos/09-Testing/Testing/src/app/demos/samples/component-testingmodule/testing-module/testing-module.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
+import { foodLoadData } from '../../foodService/food.service.mocks';
 import { FoodItem } from '../../service-http-injection/food.model';
 
 @Component({
@@ -7,12 +8,7 @@ import { FoodItem } from '../../service-http-injection/food.model';
   styleUrls: ['./testing-module.component.scss'],
 })
 export class TestingModuleComponent {
-  food: FoodItem[] = [
-    { name: 'Pad Thai', rating: 5 },
-    { name: 'Butter Chicken', rating: 5 },
-    { name: 'Cannelloni', rating: 4 },
-    { name: 'Cordon Bleu', rating: 2 },
-  ];
+  food: FoodItem[] = foodLoadData;
 
   constructor() {}
