+++ Demos/09-Testing/Testing/src/app/demos/samples/util/voucher.model.ts
