+++ Demos/09-Testing/Testing/src/app/home/home.component.html
@@ -1,12 +1,12 @@
 <mat-card>
-  <mat-card-header>
+  <mat-card-header >
     <mat-card-title>{{ title }}</mat-card-title>
   </mat-card-header>
   <mat-card-content fxLayout="row" fxLayoutAlign="center center">
     <img src="assets/images/testing.png" alt="">
   </mat-card-content>
   <mat-card-actions>
-    <button mat-button routerLink="/demos">Proceed</button>
+    <button mat-button routerLink="/demos" >Proceed</button>
   </mat-card-actions>
 </mat-card>
